+++ src/SearchBar.jsx
@@ -1,17 +1,23 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { Icon, Input } from "material-ui";
+import { Icon, IconButton, Input } from "material-ui";
 import { withStyles } from "material-ui/styles";
 import { grey } from "material-ui/colors";
 import debounce from "lodash.debounce";
 
 const DEBOUNCE_DURATION = 300;
 
-const styles = {
+const styles = theme => ({
   searchBar: {
     height: 64,
     display: "flex",
-    alignItems: "center"
+    alignItems: "center",
+    paddingLeft: theme.spacing.unit * 3,
+    paddingRight: theme.spacing.unit * 3,
+    borderBottom: "2px #D8D8D8 solid"
+  },
+  hidden: {
+    display: "none"
   },
   icon: {
     marginLeft: 8,
@@ -24,24 +30,42 @@ const styles = {
     color: grey[800],
     fontSize: 24
   }
-};
+});
 
 class SearchBar extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { query: "" };
+  }
+
+  handleChange = query => {
+    this.setState({ query });
+    this.debouncedOnChange(query);
+  };
+
   debouncedOnChange = debounce(this.props.onChange, this.props.debounce);
 
   render() {
-    const { classes } = this.props;
+    const { classes, isHidden } = this.props;
+    const { query } = this.state;
     return (
-      <div className={classes.searchBar}>
+      <div className={`${classes.searchBar} ${isHidden && classes.hidden}`}>
         <Input
           fullWidth
+          disableUnderline
+          classes={{ root: classes.input }}
           id="search"
           placeholder="Search"
-          type="search"
-          onChange={event => this.debouncedOnChange(event.target.value)}
+          onChange={event => this.handleChange(event.target.value)}
           startAdornment={<Icon className={classes.icon}>search</Icon>}
-          classes={{ root: classes.input }}
+          type="search"
+          value={query}
         />
+        {query && (
+          <IconButton onClick={() => this.handleChange("")}>
+            <Icon>clear</Icon>
+          </IconButton>
+        )}
       </div>
     );
   }
@@ -54,11 +78,13 @@ SearchBar.propTypes = {
     searchBar: PropTypes.string
   }).isRequired,
   debounce: PropTypes.number,
+  isHidden: PropTypes.bool,
   onChange: PropTypes.func.isRequired
 };
 
 SearchBar.defaultProps = {
-  debounce: DEBOUNCE_DURATION
+  debounce: DEBOUNCE_DURATION,
+  isHidden: false
 };
 
 export default withStyles(styles)(SearchBar);