+++ src/index.js
@@ -1,3 +1,5 @@
+export ActionBar from "./ActionBar";
+export ActionButton from "./ActionButton";
 export AutoSave from "./AutoSave";
 export AutoTable from "./AutoTable";
 export BaseTable from "./BaseTable";